refactor: Use the similar credit pattern of utilities to allow customers to create their own video players if necessary
