+++ index.js
@@ -17,12 +17,14 @@ import Separator from "./src/components/separator";
 import Stack from "./src/components/stack";
 import Video from "./src/components/video";
 import formatCredits from "./src/utils/format-credits";
+import formatPowaVideoEmbed from "./src/utils/format-powa-video-embed";
 
 export {
 	Attribution,
 	Button,
 	Date,
 	formatCredits,
+	formatPowaVideoEmbed,
 	Heading,
 	HeadingSection,
 	Icon,
